+++ src/main/kotlin/domain/Computer.kt
@@ -1,4 +1,4 @@
-package computer
+package domain
 
 class Computer(
     private val numbers: List<Int> = listOf()
